Removing Login / Registration
I have disabled Login / Registration as it is not implemented fully and I made the Card Selection as a 3D Scene with Numbers so we can play from 2x2 to 10x10 grid as i do not have to search for all the images for the grid.

It is very simple scoring system.

Basic Singleton is applied.
Advanced methodologies are not used in the project.
The Coupling and Referencing of Objects is pretty tight at this point of time.